requirements.txt
google-generativeai
PyYAML
python-dotenv
slack_bolt
google-search-results
py-trello
PyGithub
